clean up gravity, style fixes
