+++ src/Castro_gravity.H
@@ -8,9 +8,7 @@
 /// @param amr_ncycle       the number of subcycles at this level
 /// @param time             current time
 ///
-void construct_old_gravity(int amr_iteration, int amr_ncycle,
-                               amrex::Real time);
-
+void construct_old_gravity(int amr_iteration, int amr_ncycle, amrex::Real time);
 
 ///
 /// Construct gravitational field at new timestep
@@ -19,8 +17,7 @@ void construct_old_gravity(int amr_iteration, int amr_ncycle,
 /// @param amr_ncycle       the number of subcycles at this level
 /// @param time             current time
 ///
-    void construct_new_gravity(int amr_iteration, int amr_ncycle,
-                               amrex::Real time);
+void construct_new_gravity(int amr_iteration, int amr_ncycle, amrex::Real time);
 
 ///
 /// Construct gravitational source terms at old timestep
@@ -30,8 +27,9 @@ void construct_old_gravity(int amr_iteration, int amr_ncycle,
 /// @param time     current time
 /// @param dt       timestep
 ///
-    void construct_old_gravity_source(amrex::MultiFab& source, amrex::MultiFab& state, amrex::Real time, amrex::Real dt);
-
+void construct_old_gravity_source(amrex::MultiFab &source,
+                                  amrex::MultiFab &state, amrex::Real time,
+                                  amrex::Real dt);
 
 ///
 /// Construct gravitational source terms at new timestep
@@ -42,5 +40,7 @@ void construct_old_gravity(int amr_iteration, int amr_ncycle,
 /// @param time         current time
 /// @param dt           timestep
 ///
-    void construct_new_gravity_source(amrex::MultiFab& source, amrex::MultiFab& state_old, amrex::MultiFab& state_new, amrex::Real time, amrex::Real dt);
-
+void construct_new_gravity_source(amrex::MultiFab &source,
+                                  amrex::MultiFab &state_old,
+                                  amrex::MultiFab &state_new, amrex::Real time,
+                                  amrex::Real dt);
